% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{get_t_g}
\alias{get_t_g}
\title{get_t_g}
\usage{
get_t_g(params)
}
\arguments{
\item{params}{A numerical vector of length five. It encodes
[c(p_0, k_1, k_2, tau_1, tau_2)].}
}
\description{
get_t_g
}
