% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{get_t_n}
\alias{get_t_n}
\title{Title}
\usage{
get_t_n(params)
}
\arguments{
\item{params}{A numerical vector of length five. It encodes
[c(p_0, k_1, k_2, tau_1, tau_2)].}
}
\value{
A number
}
\description{
Title
}
