% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time-Invariant_Functions.R
\name{predictedPerformance}
\alias{predictedPerformance}
\title{predictedPerformance}
\usage{
predictedPerformance(params, training_load, day = length(training_load))
}
\arguments{
\item{params}{A length-five numerical vector giving the value of (in order)
p_0, k_1, k_2, tau_1, tau_2.}

\item{training_load}{A numerical vector containing the training load.
Can be any length.}

\item{day}{A number that specifies the length of the return vector.
Defaults to the length of `training_load`.}
}
\value{
A numerical vector the length of `day`.
}
\description{
Banister IR Time Invariant Predicted Performance
}
